+++ src/collect.js
@@ -1,0 +1,25 @@
+// @flow
+const { transduce } = require('./transduce');
+
+import type { Event, Stream } from './basics';
+
+function collect<T>(): (Stream<T>) => Promise<T[]> {
+  return transduce(
+    undefined,
+    (accumulation: T[], event: Event<T>) => {
+      if (event.error) {
+        throw event.error;
+      }
+
+      if (event.value != null) {
+        accumulation.push(event.value);
+      }
+
+      return { accumulation, done: event.done };
+    },
+    () => []);
+}
+
+module.exports = {
+  collect
+};